+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/property_info_without_constructor_extractor.php
@@ -1,8 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'property_info' => [
-        'enabled' => true,
-        'with_constructor_extractor' => false,
-    ],
-]);