+++ src/Symfony/Component/PropertyInfo/Tests/DependencyInjection/Fixtures/property_info_without_constructor_extractor.php
@@ -1,0 +1,3 @@
+<?php
+
+$container->loadFromExtension('property_info', ['with_constructor_extractor' => false]);